Move UpdateArchive class to its own file
